# docs/openapi.json
{
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "question": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "completion_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Tokens"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason"
          },
          "latency_ms": {
            "default": 0,
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "number"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Tokens"
          },
          "refused": {
            "title": "Refused",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "answer",
          "citations",
          "refused",
          "backend"
        ],
        "title": "AskResponse",
        "type": "object"
      },
      "Citation": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "title",
          "source",
          "score"
        ],
        "title": "Citation",
        "type": "object"
      },
      "EvaluationCase": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "case": {
            "minimum": 1.0,
            "title": "Case",
            "type": "integer"
          },
          "expected_answerable": {
            "title": "Expected Answerable",
            "type": "boolean"
          },
          "expected_document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Document Id"
          },
          "keyword_coverage": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Keyword Coverage",
            "type": "number"
          },
          "observed_citation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Observed Citation Ids",
            "type": "array"
          },
          "observed_refused": {
            "title": "Observed Refused",
            "type": "boolean"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "refusal_correct": {
            "title": "Refusal Correct",
            "type": "boolean"
          },
          "retrieval_correct": {
            "title": "Retrieval Correct",
            "type": "boolean"
          }
        },
        "required": [
          "case",
          "question",
          "expected_answerable",
          "observed_refused",
          "expected_document_id",
          "observed_citation_ids",
          "retrieval_correct",
          "refusal_correct",
          "keyword_coverage",
          "answer",
          "passed"
        ],
        "title": "EvaluationCase",
        "type": "object"
      },
      "EvaluationMetrics": {
        "properties": {
          "examples": {
            "minimum": 0.0,
            "title": "Examples",
            "type": "integer"
          },
          "keyword_coverage": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Keyword Coverage",
            "type": "number"
          },
          "refusal_accuracy": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Refusal Accuracy",
            "type": "number"
          },
          "retrieval_accuracy": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Retrieval Accuracy",
            "type": "number"
          }
        },
        "required": [
          "examples",
          "retrieval_accuracy",
          "refusal_accuracy",
          "keyword_coverage"
        ],
        "title": "EvaluationMetrics",
        "type": "object"
      },
      "EvaluationReport": {
        "properties": {
          "cases": {
            "items": {
              "$ref": "#/components/schemas/EvaluationCase"
            },
            "title": "Cases",
            "type": "array"
          },
          "dataset_version": {
            "title": "Dataset Version",
            "type": "string"
          },
          "failed_cases": {
            "minimum": 0.0,
            "title": "Failed Cases",
            "type": "integer"
          },
          "metrics": {
            "$ref": "#/components/schemas/EvaluationMetrics"
          },
          "passed_cases": {
            "minimum": 0.0,
            "title": "Passed Cases",
            "type": "integer"
          }
        },
        "required": [
          "dataset_version",
          "metrics",
          "passed_cases",
          "failed_cases",
          "cases"
        ],
        "title": "EvaluationReport",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PolicyDetail": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "title",
          "source",
          "excerpt",
          "text"
        ],
        "title": "PolicyDetail",
        "type": "object"
      },
      "PolicySummary": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "title",
          "source",
          "excerpt"
        ],
        "title": "PolicySummary",
        "type": "object"
      },
      "RuntimeLinks": {
        "properties": {
          "about": {
            "title": "About",
            "type": "string"
          },
          "assistant": {
            "title": "Assistant",
            "type": "string"
          },
          "evaluations": {
            "title": "Evaluations",
            "type": "string"
          },
          "openapi": {
            "title": "Openapi",
            "type": "string"
          },
          "policies": {
            "title": "Policies",
            "type": "string"
          }
        },
        "required": [
          "assistant",
          "policies",
          "evaluations",
          "about",
          "openapi"
        ],
        "title": "RuntimeLinks",
        "type": "object"
      },
      "RuntimeMetadata": {
        "properties": {
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "document_count": {
            "minimum": 0.0,
            "title": "Document Count",
            "type": "integer"
          },
          "evaluation_count": {
            "minimum": 0.0,
            "title": "Evaluation Count",
            "type": "integer"
          },
          "links": {
            "$ref": "#/components/schemas/RuntimeLinks"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "version",
          "backend",
          "document_count",
          "evaluation_count",
          "capabilities",
          "links"
        ],
        "title": "RuntimeMetadata",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Verified Support Assistant maintainers"
    },
    "description": "Citation-first support answers backed by a versioned policy corpus, with explicit safe refusal when evidence or authority is missing.",
    "license": {
      "name": "MIT"
    },
    "title": "Verified Support Assistant",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/ask": {
      "post": {
        "operationId": "ask_api_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/ask/stream": {
      "post": {
        "operationId": "ask_stream_api_ask_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask Stream",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/evaluations/latest": {
      "get": {
        "operationId": "latest_evaluation_api_evaluations_latest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationReport"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Latest Evaluation",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/api/meta": {
      "get": {
        "operationId": "metadata_api_meta_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeMetadata"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Metadata",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/policies": {
      "get": {
        "operationId": "policy_catalog_api_policies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PolicySummary"
                  },
                  "title": "Response Policy Catalog Api Policies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Policy Catalog",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/policies/{document_id}": {
      "get": {
        "operationId": "policy_detail_api_policies__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Policy Detail",
        "tags": [
          "discovery"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "operations"
        ]
      }
    }
  },
  "tags": [
    {
      "description": "Question answering with structured citations and streaming progress.",
      "name": "assistant"
    },
    {
      "description": "Runtime metadata and policy discovery.",
      "name": "discovery"
    },
    {
      "description": "Inspectable deterministic quality results.",
      "name": "evaluation"
    },
    {
      "description": "Health and readiness information.",
      "name": "operations"
    }
  ]
}